name holder replaced in Gene Page
