+++ src/components/GenePage.css
@@ -19,19 +19,6 @@
     padding-right: 0;
 }
 
-#gene_name_box{
-    position: relative;
-    margin-top: 0px;
-    left: 0;
-    top: 0;
-    float: left;
-    min-width: 100%;
-    margin-left: 0px;
-    padding-left: 0;
-    margin-right: 0;
-    padding-right: 0;
-}
-
 .codeCard{
     border-radius: 8px;
     box-shadow: 0px 0px 8px #ccc;
@@ -113,19 +100,4 @@
     max-width: 100%;
 }
 
-.title_tag, .gene_name{
-    margin: 0;
-    padding: 0;
-}
-
-.subtitle_tag, .subtitle_content{
-    color: gray;
-    font-size:medium;
-    margin: 0;
-    padding: 0;
-}
-
-.title_tag{
-    color: gray;
-}
