+++ src/components/NameHeaderHolder.css
@@ -1,0 +1,28 @@
+#name_box{
+    position: relative;
+    margin-top: 0px;
+    left: 0;
+    top: 0;
+    float: left;
+    min-width: 100%;
+    margin-left: 0px;
+    padding-left: 0;
+    margin-right: 0;
+    padding-right: 0;
+}
+
+.title_tag, .gene_name{
+    margin: 0;
+    padding: 0;
+}
+
+.subtitle_tag, .subtitle_content{
+    color: gray;
+    font-size:medium;
+    margin: 0;
+    padding: 0;
+}
+
+.title_tag{
+    color: gray;
+}
